+++ implementacion-java/src/position/MeasureUnit.java
@@ -2,5 +2,5 @@ package position;
 
 public interface MeasureUnit {
     //prerrequisito: El parámetro debe expresar una distancia en kilómetros
-    public int convert(int kms);
+    public double convert(double kms);
 }
