+++ implementacion-java/src/position/Meters.java
@@ -2,7 +2,7 @@ package position;
 
 public class Meters implements MeasureUnit {
     @Override
-    public int convert(int kms) {
+    public double convert(double kms) {
         return kms * 1000;
     }
 }
